Basic Light/Dark Mode via OS Preference Setting (#21)
* Initial Proof-of-Concept

This gets a basic working version of the blog using light/dark modes
based on system preference. I've only applied it to the Complete Guide
to map, as I'm planning to have another commit that'll implement a
little script to generate all of the `index.html` files from the
template.

Something like `npm run new-post`, which will loop through
all of the directories and inject a freshly-built `index.html` file.

I could then create a new post by making a new directory for it and then
running `npn run new-post`, which would both create the new
`index.html` file and standardize all of the existing ones, in case
anything had changed in the template.

Long-term, something like `npm run new -- "Some Post Name"` could be an
option, and it would build the directory path automatically... but I
don't need that.

* Format code

Removed formatting on Markdown files, as I don't want all of the code
blocks in the posts formatted by Prettier. I've written them as-is for
a reason.

* Just copy/paste everything manually